sesson 58

+++ Modules/Mlk/Comment/Http/Controllers/Admin/CommentController.php
@@ -8,6 +8,8 @@ use Mlk\Comment\Repositories\CommentRepo;
 
 class CommentController extends Controller
 {
+    private string $class = Comment::class;
+
     public CommentRepo $repo;
 
     public function __construct(CommentRepo $commentRepo)
@@ -17,12 +19,15 @@ class CommentController extends Controller
 
     public function index()
     {
+        $this->authorize('manage', $this->class);
         $comments = $this->repo->index()->paginate(10);
+
         return view('Comment::Admin.index', compact('comments'));
     }
 
     public function destroy($id)
     {
+        $this->authorize('manage', $this->class);
         $this->repo->delete($id);
 
         toast('حذف کامنت','success');
@@ -31,6 +36,7 @@ class CommentController extends Controller
 
     public function active($id)
     {
+        $this->authorize('manage', $this->class);
         $this->repo->changeStatus($id, Comment::STATUS_ACTIVE);
 
         alert()->success('فعال کردن وضعیت کامنت', 'عملیات با موفقیت انجام شد')->persistent('باشه');
@@ -39,6 +45,7 @@ class CommentController extends Controller
 
     public function inactive($id)
     {
+        $this->authorize('manage', $this->class);
         $this->repo->changeStatus($id, Comment::STATUS_INACTIVE);
 
         alert()->success('غیر فعال کردن وضعیت کامنت', 'عملیات با موفقیت انجام شد')->autoClose(5000);
